Assuming any prefix ending in .geojson is a file
